+++ src/index.ts
@@ -1,6 +1,6 @@
 export { builders } from './builders'
 export { is } from './is'
-export type { NodePath } from './nodepath'
+export type { NodePath, NodePathT } from './nodepath'
 export type { Scope, Label } from './scope'
 export type { BindingKind, BindingPathT, Binding, GlobalBinding } from './binding'
 export { traverse, VisitorFn, Visitor, Visitors, VisitorContext } from './traverse'